用户页面，和三个列表数据的呈现

+++ app/src/main/java/com/qxy/NoError/user/net/ListResponseData.java
@@ -19,7 +19,7 @@ public class ListResponseData<T> {
 
         public List<T> list;
         public Integer total;
-        public Integer cursor;
+        public Long cursor;
         public Integer error_code;
         public String description;
         public Boolean has_more;
